remove out of date fine resolution loader
