+++ external/loaders/tests/test__fine_resolution_budget.py
@@ -1,276 +1,5 @@
-import pytest
 import numpy as np
-import xarray as xr
-from vcm import safe
-from loaders.mappers._fine_resolution_budget import _convergence, FineResolutionSources
-
-coords = [(["x"], [1.0]), (["pfull"], [1.0, 0.0])]
-budget_ds_input = xr.Dataset(
-    dict(
-        T=xr.DataArray([[270.0, 280.0]], coords, ["x", "pfull"], attrs={"units": "K"}),
-        t_dt_phys_coarse=xr.DataArray(
-            [[0.1, 0.2]], coords, ["x", "pfull"], attrs={"units": "K/s"}
-        ),
-        t_dt_fv_sat_adj_coarse=xr.DataArray(
-            [[0.2, 0.3]], coords, ["x", "pfull"], attrs={"units": "K/s"}
-        ),
-        t_dt_nudge_coarse=xr.DataArray(
-            [[-0.1, 0.0]], coords, ["x", "pfull"], attrs={"units": "K/s"}
-        ),
-        eddy_flux_vulcan_omega_temp=xr.DataArray(
-            [[-0.1, 0.0]], coords, ["x", "pfull"], attrs={"units": "K Pa/s"}
-        ),
-        T_vulcan_omega_coarse=xr.DataArray(
-            [[-0.1, 0.0]], coords, ["x", "pfull"], attrs={"units": "K Pa/s"}
-        ),
-        T_storage=xr.DataArray(
-            [[-0.1, 0.0]], coords, ["x", "pfull"], attrs={"units": "K/s"}
-        ),
-        sphum=xr.DataArray(
-            [[1.0e-3, 2.0e-3]], coords, ["x", "pfull"], attrs={"units": "kg/kg"}
-        ),
-        qv_dt_phys_coarse=xr.DataArray(
-            [[1.0e-6, 2.0e-6]], coords, ["x", "pfull"], attrs={"units": "kg/kg/s"}
-        ),
-        qv_dt_fv_sat_adj_coarse=xr.DataArray(
-            [[2.0e-6, 3.0e-6]], coords, ["x", "pfull"], attrs={"units": "kg/kg/s"}
-        ),
-        sphum_storage=xr.DataArray(
-            [[2.0e-6, 3.0e-6]], coords, ["x", "pfull"], attrs={"units": "kg/kg/s"}
-        ),
-        eddy_flux_vulcan_omega_sphum=xr.DataArray(
-            [[-1.0e-6, 0.0]], coords, ["x", "pfull"], attrs={"units": "kg Pa/kg/s"}
-        ),
-        sphum_vulcan_omega_coarse=xr.DataArray(
-            [[-1.0e-6, 0.0]], coords, ["x", "pfull"], attrs={"units": "kg Pa/kg/s"}
-        ),
-        vulcan_omega_coarse=xr.DataArray(
-            [[-1.0e-6, 0.0]], coords, ["x", "pfull"], attrs={"units": "Pa/s"}
-        ),
-        delp=xr.DataArray(
-            [[200.0, 100.0]], coords, ["x", "pfull"], attrs={"units": "Pa"}
-        ),
-    )
-)
-
-
-budget_ds = xr.Dataset(
-    dict(
-        air_temperature=xr.DataArray(
-            [270.0], [(["pfull"], [1.0])], ["pfull"], attrs={"units": "K"}
-        ),
-        air_temperature_physics=xr.DataArray(
-            [0.1], [(["pfull"], [1.0])], ["pfull"], attrs={"units": "K/s"}
-        ),
-        air_temperature_saturation_adjustment=xr.DataArray(
-            [0.2], [(["pfull"], [1.0])], ["pfull"], attrs={"units": "K/s"}
-        ),
-        air_temperature_convergence=xr.DataArray(
-            [-0.1], [(["pfull"], [1.0])], ["pfull"], attrs={"units": "K/s"}
-        ),
-        specific_humidity=xr.DataArray(
-            [1.0e-3], [(["pfull"], [1.0])], ["pfull"], attrs={"units": "kg/kg"}
-        ),
-        specific_humidity_physics=xr.DataArray(
-            [1.0e-6], [(["pfull"], [1.0])], ["pfull"], attrs={"units": "kg/kg/s"}
-        ),
-        specific_humidity_saturation_adjustment=xr.DataArray(
-            [2.0e-6], [(["pfull"], [1.0])], ["pfull"], attrs={"units": "kg/kg/s"}
-        ),
-        specific_humidity_convergence=xr.DataArray(
-            [-1.0e-6], [(["pfull"], [1.0])], ["pfull"], attrs={"units": "kg/kg/s"}
-        ),
-    )
-)
-apparent_source_terms = ["physics", "saturation_adjustment", "convergence"]
-
-
-@pytest.mark.parametrize(
-    "ds, variable_name, apparent_source_name, apparent_source_terms, expected",
-    [
-        pytest.param(
-            budget_ds,
-            "air_temperature",
-            "dQ1",
-            apparent_source_terms,
-            budget_ds.assign(
-                {
-                    "dQ1": xr.DataArray(
-                        [0.2],
-                        [(["pfull"], [1.0])],
-                        ["pfull"],
-                        attrs={
-                            "name": "apparent source of air_temperature",
-                            "units": "K/s",
-                        },
-                    )
-                }
-            ),
-            id="base case",
-        ),
-        pytest.param(
-            budget_ds,
-            "air_temperature",
-            "dQ1",
-            ["physics", "saturation_adjustment"],
-            budget_ds.assign(
-                {
-                    "dQ1": xr.DataArray(
-                        [0.3],
-                        [(["pfull"], [1.0])],
-                        ["pfull"],
-                        attrs={
-                            "name": "apparent source of air_temperature",
-                            "units": "K/s",
-                        },
-                    )
-                }
-            ),
-            id="no convergence",
-        ),
-        pytest.param(
-            budget_ds,
-            "air_temperature",
-            "dQ1",
-            [],
-            budget_ds.assign(
-                {
-                    "dQ1": xr.DataArray(
-                        [0.3],
-                        [(["pfull"], [1.0])],
-                        ["pfull"],
-                        attrs={
-                            "name": "apparent source of air_temperature",
-                            "units": "K/s",
-                        },
-                    )
-                }
-            ),
-            id="empty list",
-            marks=pytest.mark.xfail,
-        ),
-    ],
-)
-def test__insert_budget_dQ(
-    ds, variable_name, apparent_source_name, apparent_source_terms, expected
-):
-    output = FineResolutionSources._insert_budget_dQ(
-        ds, variable_name, apparent_source_name, apparent_source_terms,
-    )
-    xr.testing.assert_allclose(output["dQ1"], expected["dQ1"])
-    assert output["dQ1"].attrs == expected["dQ1"].attrs
-
-
-@pytest.mark.parametrize(
-    "ds, variable_name, apparent_source_name, expected",
-    [
-        pytest.param(
-            budget_ds,
-            "air_temperature",
-            "pQ1",
-            budget_ds.assign(
-                {
-                    "pQ1": xr.DataArray(
-                        [0.0],
-                        [(["pfull"], [1.0])],
-                        ["pfull"],
-                        attrs={
-                            "name": "coarse-res physics tendency of air_temperature",
-                            "units": "K/s",
-                        },
-                    )
-                }
-            ),
-            id="base case",
-        ),
-        pytest.param(
-            xr.Dataset(
-                {
-                    "air_temperature": xr.DataArray(
-                        [270.0], [(["pfull"], [1.0])], ["pfull"]
-                    )
-                }
-            ),
-            "air_temperature",
-            "pQ1",
-            budget_ds.assign(
-                {
-                    "pQ1": xr.DataArray(
-                        [0.0],
-                        [(["pfull"], [1.0])],
-                        ["pfull"],
-                        attrs={
-                            "name": "coarse-res physics tendency of air_temperature"
-                        },
-                    )
-                }
-            ),
-            id="no units",
-        ),
-        pytest.param(
-            budget_ds,
-            "air_temperature",
-            "pQ1",
-            budget_ds.assign(
-                {
-                    "pQ1": xr.DataArray(
-                        [0.0],
-                        [(["pfull"], [1.0])],
-                        ["pfull"],
-                        attrs={
-                            "name": "coarse-res physics tendency of air_temperature",
-                            "units": "K",
-                        },
-                    )
-                }
-            ),
-            id="wrong units",
-            marks=pytest.mark.xfail,
-        ),
-    ],
-)
-def test__insert_budget_pQ(ds, variable_name, apparent_source_name, expected):
-    output = FineResolutionSources._insert_budget_pQ(
-        ds, variable_name, apparent_source_name
-    )
-    xr.testing.assert_allclose(output["pQ1"], expected["pQ1"])
-    assert output["pQ1"].attrs == expected["pQ1"].attrs
-
-
-@pytest.fixture
-def fine_res_mapper():
-    return {"20160901.001500": budget_ds_input}
-
-
-def test_FineResolutionSources(fine_res_mapper):
-    fine_res_source_mapper = FineResolutionSources(
-        fine_res_mapper, dim_order=("x", "pfull")
-    )
-    source_ds = fine_res_source_mapper["20160901.001500"]
-    safe.get_variables(
-        source_ds,
-        [
-            "dQ1",
-            "dQ2",
-            "pQ1",
-            "pQ2",
-            "air_temperature",
-            "specific_humidity",
-            "air_temperature_saturation_adjustment",
-            "air_temperature_nudging",
-            "air_temperature_physics",
-            "air_temperature_unresolved_flux",
-            "air_temperature_total_resolved_flux",
-            "air_temperature_storage",
-            "specific_humidity",
-            "specific_humidity_saturation_adjustment",
-            "specific_humidity_physics",
-            "specific_humidity_unresolved_flux",
-            "specific_humidity_total_resolved_flux",
-            "specific_humidity_storage",
-            "omega",
-        ],
-    )
+from loaders.mappers._hybrid import _convergence
 
 
 def test__convergence_constant():
